auto_water_system/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.0.2
django-cors-headers==4.6.0
django-environ==0.11.2
django-fsm==1.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
pillow==11.0.0
pluggy==1.5.0
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.0.6
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.2
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.18.3
